1. Cleared the null pointer exceptions 2. Updated the code for changing the name to ID 3. Created new folder for holding all the JSPs of the concept collections

+++ trunk/Quadriga/src/main/java/edu/asu/spring/quadriga/web/ConceptcollectionController.java
@@ -189,12 +189,14 @@ public class ConceptcollectionController {
 		
 		
 		String[] values= req.getParameterValues("selected");
+		if(values!=null){
 		for(String id : values)
 		{
 			conceptControllerManager.deleteItem(id,concept.getName());
-			con = conFact.createConceptObject();con.setName(id);
+			con = conFact.createConceptObject();con.setId(id);
 			concept.getItems().remove(concept.getItems().indexOf(con));
 			
+		}
 		}
 		
 		conceptControllerManager.getCollectionDetails(concept);
@@ -205,6 +207,7 @@ public class ConceptcollectionController {
 	@RequestMapping(value = "auth/conceptcollections/updateitems", method = RequestMethod.POST)
 	public String conceptUpdateHandler(HttpServletRequest req, ModelMap model) throws SQLException{
 		String[] values= req.getParameterValues("selected");
+		if(values!=null)
 		conceptControllerManager.update(values,concept);
 		
 		return "redirect:/auth/conceptdetails";
